requirements.txt
streamlit==1.25.0
rembg==2.0.30
pillow
ultralytics
opencv-python-headless
